+++ api/plugins/client.ts
@@ -2,7 +2,7 @@ import { appendHeader } from "h3";
 import { type FetchOptions, FetchError } from "ofetch";
 import { splitCookiesString } from "set-cookie-parser";
 import ApiError from "../models/ApiError";
-import type User from "../models/User";
+import type User from "../../interfaces/User/User";
 import type { ApiServiceContainer } from "../services/ApiServiceContainer";
 import ApplicationService from "../services/ApplicationService";
 import AuthenticationService from "../services/AuthenticationService";
