+++ services/api.ts
@@ -1,14 +1,13 @@
 import axios, { type AxiosInstance } from 'axios';
 import { useUserStore } from '@/stores/User';
 
-// const defaultUrl: string = "http://gyofood.test/api/qr-menu/v1";
-const defaultUrl: string = "https://app.myrest.pe/api/qr-menu/v1";
+const defaultUrl: string = process.env.NUXT_PUBLIC_API_BASE_URL??'';
 
-const api: AxiosInstance = axios.create({
+const apiAxios: AxiosInstance = axios.create({
   baseURL: defaultUrl,
 });
 
-api.interceptors.request.use((config) => {
+apiAxios.interceptors.request.use((config) => {
   const userStore = useUserStore();
   const token: string | null = userStore.getToken;
 
@@ -22,4 +21,4 @@ api.interceptors.request.use((config) => {
   return config;
 }, (error) => Promise.reject(error));
 
-export default api;
+export default apiAxios;
